move spnego_fat sshd calls to fattest.simplicity
