+++ dev/com.ibm.ws.security.spnego.fat.common/bnd.bnd
@@ -37,7 +37,4 @@ test.project: true
     com.ibm.ws.security.token.s4u2;version=latest,\
     com.ibm.ws.webcontainer.security;version=latest,\
     com.ibm.ws.webcontainer.security_test.servlets;version=latest,\
-    org.apache.httpcomponents:httpcore;version=4.1.2,\
-    org.apache.sshd:sshd-common;version=2.9.2,\
-    org.apache.sshd:sshd-core;version=2.9.2,\
-    org.apache.sshd:sshd-scp;version=2.9.2
+    org.apache.httpcomponents:httpcore;version=4.1.2
