+++ dev/fattest.simplicity/src/componenttest/topology/utils/SshUtils.java
@@ -10,29 +10,26 @@
  * Contributors:
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
-package com.ibm.ws.security.spnego.fat.config;
+package componenttest.topology.utils;
 
+import java.io.IOException;
 import java.net.InetAddress;
 import java.net.UnknownHostException;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Random;
+import java.util.concurrent.TimeUnit;
 
-//import org.apache.sshd.client.SshClient;
-//import org.apache.sshd.client.session.ClientSession;
+import org.apache.sshd.client.SshClient;
+import org.apache.sshd.client.session.ClientSession;
 
-import com.ibm.websphere.simplicity.ConnectionInfo;
 import com.ibm.websphere.simplicity.Machine;
 import com.ibm.websphere.simplicity.log.Log;
-
-import componenttest.topology.utils.ExternalTestService;
-import componenttest.topology.utils.SshUtils;
 
 /**
  * Purpose: This suite collects and runs all known good test suites.
  */
-public class InitClass {
-    private static final Class<?> c = InitClass.class;
+public class SshUtils {
+    private static final Class<?> c = SshUtils.class;
     /**
      * Active Directory appears to allow user names with a max of 20 characters. We add "_http" or
      * "_httpsn" to the end of the host name to create the user name, so the host name we use for
@@ -103,91 +100,31 @@ public class InitClass {
     public static String serverShortHostName = null;
     public static boolean randomizeHostName = false;
 
-    public static void getKDCInfoFromConsul() throws Exception {
-        String thisMethod = "getKDCInfoFromConsul";
-        Log.info(c, thisMethod, "Getting KDCs from Consul.....");
-        List<ExternalTestService> services = null;
-
-        try {
-            //obtaining kdcp and kdc_r information
-            services = CommonTest.getKDCServices(2, SPNEGOConstants.KDC_HOST_FROM_CONSUL);
-            KDC_HOSTNAME = services.get(0).getAddress();
-            KDC_USER = services.get(0).getProperties().get(SPNEGOConstants.MS_KDC_USER_CONSUL);
-            KDC_USER_PWD = services.get(0).getProperties().get(SPNEGOConstants.MS_KDC_USER_PASSWORD_CONSUL);
-            KDC_REALM = services.get(0).getProperties().get(SPNEGOConstants.KDC_REALM_FROM_CONSUL);
-            KDC_HOST_SHORTNAME = services.get(0).getProperties().get(SPNEGOConstants.KDC_SHORTNAME_FROM_CONSUL);
-            KRB5_CONF = services.get(0).getProperties().get(SPNEGOConstants.KRB5_CONF_FROM_CONSUL);
-            Z_USER = services.get(0).getProperties().get(SPNEGOConstants.Z_USER_FROM_CONSUL);
-            FIRST_USER = services.get(0).getProperties().get(SPNEGOConstants.FIRST_USER_FROM_CONSUL);
-            SECOND_USER = services.get(0).getProperties().get(SPNEGOConstants.SECOND_USER_FROM_CONSUL);
-            USER_PWD = services.get(0).getProperties().get(SPNEGOConstants.USER_PWD_FROM_CONSUL);
-            Z_USER_PWD = services.get(0).getProperties().get(SPNEGOConstants.USER0_PWD_FROM_CONSUL);
-
-            ConnectionInfo connInfo = new ConnectionInfo(KDC_HOSTNAME, InitClass.KDC_USER, InitClass.KDC_USER_PWD);
-            Machine kdcMachine = Machine.getMachine(connInfo);
-
+    /**
+     * @param  thisMethod
+     * @param  kdcMachine
+     * @throws Exception
+     * @throws InterruptedException
+     */
+    public static void establishConnectionToKDC(String thisMethod, Machine kdcMachine) throws Exception, InterruptedException {
+        for (int i = 1; i <= 3; i++) {
             try {
-                Log.info(c, thisMethod, "Testing connection to KDC: " + KDC_HOST_SHORTNAME);
-                SshUtils.establishConnectionToKDC(thisMethod, kdcMachine);
+                SshClient sshClient = getSshClient();
+                try {
+                    getSshSession(sshClient, kdcMachine);
+                } finally {
+                    sshClient.stop();
+                }
+                Log.info(c, thisMethod, "sshUtil KDC connection succeeded after " + i + " attempt(s)");
+                break;
             } catch (Exception e) {
-                String failedKdcShortName = KDC_HOST_SHORTNAME;
-                KDC_HOSTNAME = services.get(1).getAddress();
-                KDC_USER = services.get(1).getProperties().get(SPNEGOConstants.MS_KDC_USER_CONSUL);
-                KDC_USER_PWD = services.get(1).getProperties().get(SPNEGOConstants.MS_KDC_USER_PASSWORD_CONSUL);
-                KDC_REALM = services.get(1).getProperties().get(SPNEGOConstants.KDC_REALM_FROM_CONSUL);
-                KDC_HOST_SHORTNAME = services.get(1).getProperties().get(SPNEGOConstants.KDC_SHORTNAME_FROM_CONSUL);
-                KRB5_CONF = services.get(1).getProperties().get(SPNEGOConstants.KRB5_CONF_FROM_CONSUL);
-                Z_USER = services.get(1).getProperties().get(SPNEGOConstants.Z_USER_FROM_CONSUL);
-                FIRST_USER = services.get(1).getProperties().get(SPNEGOConstants.FIRST_USER_FROM_CONSUL);
-                SECOND_USER = services.get(1).getProperties().get(SPNEGOConstants.SECOND_USER_FROM_CONSUL);
-                USER_PWD = services.get(1).getProperties().get(SPNEGOConstants.USER_PWD_FROM_CONSUL);
-                Z_USER_PWD = services.get(1).getProperties().get(SPNEGOConstants.USER0_PWD_FROM_CONSUL);
-
-                Log.info(c, thisMethod, "connection to " + failedKdcShortName + " failed. Attempting failover KDC: " + KDC_HOST_SHORTNAME);
-
-                connInfo = new ConnectionInfo(KDC_HOSTNAME, InitClass.KDC_USER, InitClass.KDC_USER_PWD);
-                kdcMachine = Machine.getMachine(connInfo);
-                //establishConnectionToKDC(thisMethod, kdcMachine);
-                SshUtils.establishConnectionToKDC(thisMethod, kdcMachine);
+                if (i == 3) {
+                    Log.info(c, thisMethod, "sshUtil KDC connection still failed after retrying " + i + " attempts");
+                    throw e;
+                }
+                Thread.sleep(5000);
             }
-
-            KDCP_VAR = getKDCHostnameMask(KDC_HOSTNAME);
-
-            //obtaining kdcs and kdcs_r information
-            services = CommonTest.getKDCServices(1, SPNEGOConstants.KDC2_HOST_FROM_CONSUL);
-            KDC2_HOSTNAME = services.get(0).getAddress();
-            KDC2_USER = services.get(0).getProperties().get(SPNEGOConstants.MS_KDC_USER_REALM_2_CONSUL);
-            KDC2_USER_PWD = services.get(0).getProperties().get(SPNEGOConstants.MS_KDC_PASSWORD_REALM_2_CONSUL);
-            KDC2_REALM = services.get(0).getProperties().get(SPNEGOConstants.KDC2_REALM_FROM_CONSUL);
-
-            KDCS_VAR = getKDCHostnameMask(KDC2_HOSTNAME);
-
-            FIRST_USER_KRB5_FQN_PWD = USER_PWD;
-            FIRST_USER_PWD = USER_PWD;
-            SECOND_USER_PWD = USER_PWD;
-            SECOND_USER_KRB5_FQN_PWD = USER_PWD;
-            FQN = "@" + KDC_REALM;
-            FIRST_USER_KRB5_FQN = FIRST_USER + FQN;
-            SECOND_USER_KRB5_FQN = SECOND_USER + FQN;
-            COMMON_TOKEN_USER = FIRST_USER;
-            COMMON_TOKEN_USER_PWD = FIRST_USER_PWD;
-
-            // get canonical and short host name
-            getServerCanonicalHostName();
-            getServerShortHostName();
-
-        } catch (Exception e) {
-            Log.info(c, thisMethod, "An Exception happened while getting the KDC from consul. TESTS WILL NOT RUN");
-            e.printStackTrace();
-            throw e;
-
-        } finally {
-            CommonTest.releaseServices(services);
         }
-
-        Log.info(c, thisMethod, "We were able to retrieve the required information from consul, the test will now continue....");
-        Log.info(c, thisMethod, "The following KDC's are being used: " + KDCP_VAR + " and " + KDCS_VAR + ".");
-
     }
 
     public static String getKDCHostnameMask(String hostname) {
@@ -268,7 +205,7 @@ public class InitClass {
     /**
      * EBC test machines have long host names, create a random string host name for EBC test machine.
      *
-     * @param canonicalHostName
+     * @param  canonicalHostName
      * @return
      */
     protected static String createRandomStringHostNameForEbc(String canonicalHostName) {
@@ -294,7 +231,7 @@ public class InitClass {
     /**
      * Some test machines have long host names, create a random string host name for long named test machines.
      *
-     * @param canonicalHostName
+     * @param  canonicalHostName
      * @return
      */
     protected static String createRandomStringHostName(String canonicalHostName) {
@@ -322,9 +259,9 @@ public class InitClass {
      * Returns the short host name from the canonical host name provided. If canonicalHostName does not include
      * "ibm.com", the same value provided for canonicalHostName is returned.
      *
-     * @param canonicalHostName
-     * @param issueMsg          - Boolean indicating whether a message should be logged if the canonical host name does not
-     *                              include the IBM domain.
+     * @param  canonicalHostName
+     * @param  issueMsg          - Boolean indicating whether a message should be logged if the canonical host name does not
+     *                               include the IBM domain.
      * @return
      */
     public static String getShortHostName(String canonicalHostName, boolean issueMsg) {
@@ -354,4 +291,29 @@ public class InitClass {
         return getShortHostName(serverCanonicalHostName, true);
     }
 
+    /**
+     * Get a (started) SshClient.
+     *
+     * @return The SshClient.
+     */
+    protected static SshClient getSshClient() {
+        SshClient sshClient = SshClient.setUpDefaultClient();
+        sshClient.start();
+        return sshClient;
+    }
+
+    /**
+     * Get an SSH ClientSession to the specified machine.
+     *
+     * @param  sshClient   The SSH client.
+     * @param  machine     The machine to connect to.
+     * @return             The session.
+     * @throws IOException If there was an error getting an SSH session to the machine.
+     */
+    protected static ClientSession getSshSession(SshClient sshClient, Machine machine) throws IOException {
+        ClientSession session = sshClient.connect(machine.getUsername(), machine.getHostname(), 22).verify(30, TimeUnit.SECONDS).getSession();
+        session.addPasswordIdentity(machine.getPassword());
+        session.auth().verify(30, TimeUnit.SECONDS).isSuccess();
+        return session;
+    }
 }
